buteco_bot/requirements.txt
discord.py
requests
aiohttp
asyncio
python-dotenv
matplotlib
pillow

